clean some functions whit -Weverything and check nupi ok, set get expo ok,  pow int ok, rm decimal ok, set nan inf ok YEAH

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: vphongph <vphongph@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 13:34:32 by vphongph          #+#    #+#             */
-/*   Updated: 2019/02/25 15:50:07 by vphongph         ###   ########.fr       */
+/*   Updated: 2019/03/30 19:55:08 by vphongph         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,12 +21,12 @@ int		ft_atoi(const char *str)
 	if ((s = (*str == '-') ? -1 : 1) == -1 || *str == '+')
 		str++;
 	while (*str >= '0' && *str <= '9')
-		nb = nb * 10 - '0' + *str++;
+		nb = nb * 10 - '0' + (unsigned long long)*str++;
 	if (nb > 9223372036854775807ULL && s == 1)
 		return (-1);
 	if (nb > 9223372036854775808ULL && s == -1)
 		return (0);
-	return (nb * s);
+	return ((int)nb * s);
 }
 
 /*
